[nvim] get rid of vim2hs
This speeds up startup times by a few tens of milliseconds (pretty good)
because, apparently, has('python') is slow in Neovim. The feature that
required Python (and Python2 at that) was hpaste integration in vim2hs.
It's not like I program in Haskell at all these days, plus vim-polyglot
includes a good syntax and indent scripts for it.