[zsh] display paths shortened with `~/` in the `j` and `d` pickers
Note that I changed those functions to use `--accept-nth` flag of fzf,
which was added only in v0.60.0. Such a new version (released 9 months
ago) will not be available in mainline Ubuntu repos for a long time, so
on Ubuntu this practically necessitates the usage of ZPLG to install a
newer version of fzf.