[zsh] add a MANPAGER that also generates a ctags file for less(1)
This feature was stolen from OpenBSD's mandoc. Their viewer generates a
ctags-compatible tag file that less(1) can load for you, so that you are
able to jump to different nodes in the manpage with the `:t` command.
This makes browsing and finding stuff in long pages like bash(1) or
zshall(1) so much easier!