[scripts] move utility modules for python scripts into scripts/dotfiles/
This makes them easier to import from files in the `scripts/` directory,
without needing to add `../script-resources/` to `sys.path` first. Also,
as far as I understand, putting a subdirectory into a $PATH entry should
not create any problems, since $PATH is searched in the first place only
if a command is given without a `/` in the name -- therefore, it should
not be possible to "invoke" some script in `scripts/dotfiles/` by
typing, say, `dotfiles/common_script_utils.py` into the command line.